decomaltofloat

+++ ERP.Services.API/Entities/QuotationProductEntity.cs
@@ -17,6 +17,6 @@ public class QuotationProductEntity
 
     [Column("price")] public int Price { get; set; }
     [Column("order")] public int Order { get; set; }
-    [Column("amount")] public decimal Amount { get; set; }
-    [Column("discount")] public decimal Discount { get; set; }
+    [Column("amount")] public float Amount { get; set; }
+    [Column("discount")] public float Discount { get; set; }
 }
