+++ ERP.Services.API/Models/RequestModels/Quotation/QuotationProductResource.cs
@@ -4,6 +4,6 @@ public class QuotationProductResource
 {
     public Guid ProductId { get; set; }
     public int Quantity { get; set; }
-    public int Discount { get; set; }
+    public float Discount { get; set; }
    
 }
